+++ amplify/backend/function/yAppLambda/src/Controllers/PostController.cs
@@ -109,18 +109,19 @@ public class PostController : ControllerBase
     /// Retrieves all public posts from a user
     /// </summary>
     /// <param name="userName">The username used to find all posts created by a user.</param>
-    /// <returns>A list of public posts created by a user.</returns>
+    /// <param name="diaryEntry">If the query is for public posts or diary entries.</param>
+    /// <returns>A list of posts created by a user, either public posts or diary entries.</returns>
     [HttpGet("getPostsByUser")]
     [ProducesResponseType(typeof(List<Post>), StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
-    public async Task<ActionResult<List<Post>>> GetPostsByUser(string userName)
+    public async Task<ActionResult<List<Post>>> GetPostsByUser(string userName, bool diaryEntry)
     {
         if(string.IsNullOrEmpty(userName))
         {
             return BadRequest("username is required");
         }
 
-        var posts = await _postActions.GetPostsByUser(userName);
+        var posts = await _postActions.GetPostsByUser(userName, diaryEntry);
 
         return posts;
     }